# doc/openapi.json
{"openapi":"3.1.0","info":{"title":"Hair Therapy API Documentation","description":"API for managing users, appointments, and blogs for the Hair Therapy system.","version":"1.0.0"},"servers":[{"url":"http://localhost:9090","description":"Generated server url"}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/services/{id}":{"get":{"tags":["service-controller"],"operationId":"getServiceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Service"}}}}}},"put":{"tags":["service-controller"],"operationId":"updateService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Service"}}}}}},"delete":{"tags":["service-controller"],"operationId":"deleteService","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/blogs/{id}":{"get":{"tags":["blog-controller"],"operationId":"getBlogById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Blog"}}}}}},"put":{"tags":["blog-controller"],"operationId":"updateBlog","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Blog"}}}}}},"delete":{"tags":["blog-controller"],"operationId":"deleteBlog","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/appointments/{id}/status":{"put":{"tags":["appointment-controller"],"operationId":"updateAppointmentStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"value","in":"query","required":true,"schema":{"type":"string","enum":["PENDING","CONFIRMED","COMPLETED","CANCELLED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Appointment"}}}}}}},"/api/appointments/{id}/cancel":{"put":{"tags":["appointment-controller"],"operationId":"cancelAppointment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Appointment"}}}}}}},"/api/users/login":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/add":{"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/services":{"get":{"tags":["service-controller"],"operationId":"getAllServices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Service"}}}}}}},"post":{"tags":["service-controller"],"operationId":"addService","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Service"}}}}}}},"/api/blogs":{"get":{"tags":["blog-controller"],"operationId":"getAllBlogs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Blog"}}}}}}},"post":{"tags":["blog-controller"],"operationId":"addBlog","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Blog"}}}}}}},"/api/appointments/book":{"post":{"tags":["appointment-controller"],"operationId":"bookAppointment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppointmentDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Appointment"}}}}}}},"/api/appointments/stylist/{stylistId}":{"get":{"tags":["appointment-controller"],"operationId":"getAppointmentsByStylist","parameters":[{"name":"stylistId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}}}},"/api/appointments/client/{clientId}":{"get":{"tags":["appointment-controller"],"operationId":"getAppointmentsByClient","parameters":[{"name":"clientId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}}}},"/api/appointments/all":{"get":{"tags":["appointment-controller"],"operationId":"getAllAppointments","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Appointment"}}}}}}}}},"components":{"schemas":{"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["CLIENT","STYLIST","ADMIN"]}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["CLIENT","STYLIST","ADMIN"]}}},"ServiceDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"float"},"duration":{"type":"integer","format":"int32"},"stylistId":{"type":"integer","format":"int32"}}},"Service":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"float"},"duration":{"type":"integer","format":"int32"},"stylist":{"$ref":"#/components/schemas/User"}}},"BlogDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"authorId":{"type":"integer","format":"int32"}}},"Blog":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"author":{"$ref":"#/components/schemas/User"}}},"Appointment":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"bookingTime":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["PENDING","CONFIRMED","COMPLETED","CANCELLED"]},"client":{"$ref":"#/components/schemas/User"},"stylist":{"$ref":"#/components/schemas/User"},"service":{"$ref":"#/components/schemas/Service"}}},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"AppointmentDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"bookingTime":{"type":"string","format":"date-time"},"serviceId":{"type":"integer","format":"int32"},"stylistId":{"type":"integer","format":"int32"},"clientId":{"type":"integer","format":"int32"}}}}}}